Suppress vim_strnsave() warning
Warning is occurred at here:
https://github.com/vim/vim/blob/7306d6b1c9528dcfe89eb77aa7ef354ffa9605a8/src/vim9compile.c#L4446

Maybe, it's better to change the prototype of vim_strnsave() than adding
type cast.

+++ src/misc2.c
@@ -1287,7 +1287,7 @@ vim_strsave(char_u *string)
  * shorter.
  */
     char_u *
-vim_strnsave(char_u *string, int len)
+vim_strnsave(char_u *string, size_t len)
 {
     char_u	*p;
 
@@ -1534,7 +1534,7 @@ vim_strsave_up(char_u *string)
  * This uses ASCII lower-to-upper case translation, language independent.
  */
     char_u *
-vim_strnsave_up(char_u *string, int len)
+vim_strnsave_up(char_u *string, size_t len)
 {
     char_u *p1;
